+++ LightScript.GitHubWorkflow.psdevops.ps1
@@ -1,13 +1,10 @@
 ﻿#requires -Module PSDevOps
 Push-Location $PSScriptRoot
-
-New-GitHubWorkflow -Name "Analyze, Test, Tag, and Publish" -On Push, PullRequest, Demand -Job PowerShellStaticAnalysis, TestPowerShellOnLinux, TagReleaseAndPublish |
+Import-BuildStep -Module LightScript
+New-GitHubWorkflow -Name "Analyze, Test, Tag, and Publish" -On Push, PullRequest, Demand -Job PowerShellStaticAnalysis, TestPowerShellOnLinux, TagReleaseAndPublish, BuildModule |
     Set-Content .\.github\workflows\TestAndPublish.yml -Encoding UTF8 -PassThru
 
-New-GitHubWorkflow -Name "Run HelpOut" -On Push  -Job HelpOut |
-    Set-Content .\.github\workflows\UpdateDocs.yml -Encoding UTF8 -PassThru
-
-New-GitHubWorkflow -Name "Run EZOut" -On Push -Job RunEZOut |
-    Set-Content .\.github\workflows\RunEZOut.yml -Encoding UTF8 -PassThru
+New-GitHubWorkflow -On Issue, Demand -Job RunGitPub -Name OnIssueChanged |
+    Set-Content (Join-Path $PSScriptRoot .github\workflows\OnIssue.yml) -Encoding UTF8 -PassThru
 
 Pop-Location