+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.6:
+* Set-NanoLeaf:
+  * Fixing -SaturationIncrement (#35)
+  * Adding -BrightnessIncrement (#34)
+  * Fixing -Hue/-Saturation/-Brightness behavior to allow increment parameters to work. (#36)
+  * Making Parameter Aliases consistent with Set-HueLight (#38)
+* Now using [GitPub](https://github.com/StartAutomating/GitPub) to blog (#37).
+---
+
 ## 0.2.5:
 * Set-HueRule:  Easier conditions (Fixes #28) and plural aliases (Fixes #31)
 * Adding Rename-HueSensor (Fixes #26).